bio-circuit-ai/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
httpx>=0.27.0
qdrant-client>=1.8.0
sentence-transformers>=2.5.0
openai>=1.14.0
biopython>=1.83
beautifulsoup4>=4.12.0
lxml>=5.1.0
numpy>=1.26.0
rich>=13.7.0
tenacity>=8.2.0
python-dotenv>=1.0.0
mcp>=1.26.0